добавила контроллер для board
