+++ src/controllers/board.js
@@ -1,0 +1,71 @@
+import LoadMoreButtonComponent from "../components/load-more-button.js";
+import SortComponent from "../components/sort.js";
+import TaskComponent from "../components/task.js";
+import TaskEditComponent from "../components/task-edit.js";
+import TasksComponent from "../components/tasks.js";
+import {remove, render, replace, RenderPosition} from "../utils/render.js";
+
+
+const SHOWING_TASKS_COUNT_ON_START = 8;
+const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
+
+const renderTask = (taskListElement, task) => {
+  const replaceTaskToEdit = () => {
+    replace(taskEditComponent, taskComponent);
+  };
+
+  const replaceEditToTask = () => {
+    replace(taskComponent, taskEditComponent);
+  };
+
+  const taskComponent = new TaskComponent(task);
+  taskComponent.setEditButtonClickHandler(() => {
+    replaceTaskToEdit();
+  });
+
+  const taskEditComponent = new TaskEditComponent(task);
+  taskEditComponent.setSubmitHandler((evt) => {
+    evt.preventDefault();
+    replaceEditToTask();
+  });
+
+  render(taskListElement, taskComponent, RenderPosition.BEFOREEND);
+};
+
+export default class BoardController {
+  constructor(container) {
+    this._container = container;
+
+    this._sortComponent = new SortComponent();
+    this._tasksComponent = new TasksComponent();
+    this._loadMoreButtonComponent = new LoadMoreButtonComponent();
+  }
+
+  render(tasks) {
+    const container = this._container.getElement();
+    render(container, this._sortComponent, RenderPosition.BEFOREEND);
+    render(container, this._tasksComponent, RenderPosition.BEFOREEND);
+
+    const taskListElement = this._tasksComponent.getElement();
+
+    let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
+    tasks.slice(0, showingTasksCount)
+      .forEach((task) => {
+        renderTask(taskListElement, task);
+      });
+
+    render(container, this._loadMoreButtonComponent, RenderPosition.BEFOREEND);
+
+    this._loadMoreButtonComponent.setCLickHandler(() => {
+      const prevTasksCount = showingTasksCount;
+      showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
+
+      tasks.slice(prevTasksCount, showingTasksCount).forEach((task) => renderTask(taskListElement, task));
+
+      if (showingTasksCount >= tasks.length) {
+        remove(this._loadMoreButtonComponent);
+      }
+    });
+
+  }
+}
